+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.1.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Update object modification date if cropping was removed/updated
+  [boulch, laulaz]
 
 
 1.1.5 (2023-03-14)